requirements.txt
apache-airflow-providers-amazon
apache-airflow-providers-postgres
pandas
requests
psycopg2-binary
sqlalchemy